requirements.txt
click
requests

pandas
lxml
xlrd
pyyaml
markdown
xlwt
numpy

matplotlib
seaborn
altair
plotly
